Adds start time to rounds, events, and makeup days
Extends calendar events, rounds, and makeup days to include start time, allowing for better scheduling.

Updates the database schema and REST endpoints to incorporate and validate start times. Adds a "Youth" status to Users.

Enhances the frontend to manage and display start times for events, rounds, and makeup days in tournament admin screens and the general calendar. This ensures that event schedules are comprehensive and accurate.